Cleaning up stuff
- updated README.md
- renamed dir
- renamed manage to shell script
- linting
- mypy types
- typo fixes
- built docker containers
- ran all notebooks
