--- examples/hl-aries-credential-exchange/.gitignore
@@ -1,12 +1,0 @@
-.idea
-
-.env
-
-.code
-dist
-*.egg-info
-build
-.ipynb_checkpoints/
-__pycache__/
-
-.vscode/
