--- examples/hl-aries-credential-exchange/dockerfiles/Dockerfile.controller
@@ -1,35 +1,0 @@
-FROM continuumio/miniconda3
-
-ENV WORKSPACE /workspace
-
-ARG jupyter_port
-ENV env_jupyter_port=jupyter_port
-
-# Setup workspace environment
-RUN apt-get update && apt-get install -y gcc
-# RUN conda install -c conda-forge jupyterlab==3.0.11
-RUN conda install -c conda-forge jupyterlab==3.0.11
-
-# Install the aries_cloudcontroller
-
-
-
-RUN pip install syft==0.5.0rc1
-
-RUN pip install aries_cloudcontroller
-
-
-# Create jupyter notebook workspace
-#RUN mkdir $WORKSPACE
-WORKDIR $WORKSPACE
-
-# Make the image start the jupyter notebook
-COPY scripts/jupyter-entrypoint.sh /entrypoint.sh
-RUN chmod +x /entrypoint.sh
-
-
-
-
-
-# ENTRYPOINT ["/entrypoint.sh", "$env_jupyter_port"]
-ENTRYPOINT ["/entrypoint.sh", "8888"]
