--- examples/hl-aries-credential-exchange/scripts/startup.sh
@@ -1,4 +1,0 @@
-aca-py start \
-    -it http '0.0.0.0' "$HTTP_PORT" \
-    -e "$AGENT_ENDPOINT" "${AGENT_ENDPOINT/http/ws}" \
-    "$@"
