+++ examples/hyperledger-aries/dockerfiles/Dockerfile.agent
