+++ examples/hyperledger-aries/dockerfiles/Dockerfile.controller
@@ -1,0 +1,21 @@
+FROM continuumio/miniconda3
+
+ENV WORKSPACE /workspace
+
+# Setup workspace environment
+RUN apt-get update && apt-get install -y gcc
+RUN conda install -c conda-forge jupyterlab==3.0.11
+
+# Install the aries_cloudcontroller
+# TODO: Update to official 0.5.0 release when available
+RUN pip install syft==0.5.0rc1
+RUN pip install aries_cloudcontroller
+
+# Create jupyter notebook workspace
+WORKDIR $WORKSPACE
+
+# Make the image start the jupyter notebook
+COPY scripts/jupyter-entrypoint.sh /entrypoint.sh
+RUN chmod +x /entrypoint.sh
+
+ENTRYPOINT ["/entrypoint.sh", "8888"]
