+++ examples/hyperledger-aries/playground/data-owner/data-owner-example.env
@@ -1,0 +1,79 @@
+# These are dummy values that can be used to get started quickly
+# Replace these with real values when deploying to a real environment
+
+# For postgres DB which hosts the wallets
+WALLET_DB_HOST=data-owner-wallet-db
+WALLET_DB_PORT=5432
+WALLET_DB_USER=postgres
+WALLET_DB_PASS=dbpass
+WALLET_DB_ADMIN_USER=postgres
+WALLET_DB_ADMIN_PASS=dbpass
+
+# These values are only used locally by docker-compose to set up the DB container - should match values above
+POSTGRES_PASSWORD=dbpass
+POSTGRES_USER=postgres
+POSTGRES_DB=data-owner_wallet
+
+# for tails server local use only
+# GENESIS_URL=http://localhost:9000/genesis
+# STORAGE_PATH=/tmp/tails-files
+# LOG_LEVEL=INFO
+# TAILS_URL=http://tails-server:6543
+
+# for aca-py agent
+HTTP_PORT=4020
+WEBHOOK_PORT=4010
+ADMIN_PORT=4021
+ADMIN_URL=http://data-owner-agent:4021
+ACAPY_OUTBOUND_TRANSPORT=http
+ACAPY_ADMIN=[0.0.0.0,4021]
+
+# This is set in the startup.sh script to include the ngrok endpoint
+ACAPY_ENDPOINT=http://data-owner-agent:4020
+
+ACAPY_WEBHOOK_URL=http://data-owner-business-logic:4010
+ACAPY_ADMIN_API_KEY=dataownerssecretapikey
+ACAPY_LABEL=Data Owner
+ACAPY_WALLET_NAME=dataowner_Name
+ACAPY_WALLET_KEY=data-owner_key
+ACAPY_WALLET_TYPE=indy
+ACAPY_WALLET_STORAGE_TYPE=postgres_storage
+
+## Optionally Specify a Seed
+# ACAPY_WALLET_SEED=0000000000000000000000000000Any1
+
+ACAPY_WALLET_STORAGE_CONFIG={"url":"data-owner-wallet-db:5432","wallet_scheme":"MultiWalletSingleTable"}
+ACAPY_WALLET_STORAGE_CREDS={"account":"postgres","password":"dbpass","admin_account":"postgres","admin_password":"dbpass"}
+
+ACAPY_LOG_LEVEL=info
+ACAPY_AUTO_PROVISION=true
+
+
+## Local Network
+# ACAPY_GENESIS_FILE=/home/indy/resources/pool_transactions_genesis_local_dev
+
+# Sovrin StagingNet
+ACAPY_GENESIS_URL=https://raw.githubusercontent.com/sovrin-foundation/sovrin/master/sovrin/pool_transactions_sandbox_genesis
+
+
+# Multi-tenant Configuration
+# ACAPY_MULTITENANT=true
+# ACAPY_MULTITENANT_ADMIN=true
+# ACAPY_MULTITENANT_JWT_SECRET=jwtSecret
+
+# Name of ngrok container if exposing agent endpoint over ngrok
+NGROK_NAME= ngrok-data-owner
+
+# Optional Helper Configurations - See https://github.com/hyperledger/aries-cloudagent-python/blob/main/aries_cloudagent/config/argparse.py
+ACAPY_AUTO_ACCEPT_INVITES=true
+ACAPY_AUTO_ACCEPT_REQUESTS=true
+ACAPY_AUTO_PING_CONNECTION=true
+ACAPY_AUTO_RESPOND_MESSAGES=false
+ACAPY_AUTO_RESPOND_CREDENTIAL_PROPOSAL=true
+ACAPY_AUTO_RESPOND_CREDENTIAL_OFFER=true
+ACAPY_AUTO_RESPOND_CREDENTIAL_REQUEST=true
+ACAPY_AUTO_RESPOND_PRESENTATION_PROPOSAL=true
+ACAPY_AUTO_RESPOND_PRESENTATION_REQUEST=true
+ACAPY_AUTO_STORE_CREDENTIAL=true
+ACAPY_AUTO_VERIFY_PRESENTATION=false
+ACAPY_PRESERVE_EXCHANGE_RECORDS=true
