+++ examples/hyperledger-aries/playground/data-scientist/data-scientist-example.env
