+++ examples/hyperledger-aries/playground/om-authority/om-authority-example.env
