+++ examples/hyperledger-aries/resources/init.sql
