+++ examples/hyperledger-aries/scripts/startup.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+aca-py start \
+-it http '0.0.0.0' "$HTTP_PORT" \
+-e "$AGENT_ENDPOINT" "${AGENT_ENDPOINT/http/ws}" \
+"$@"
